pyproject.toml
[tool.poetry]
name = "runpod-ollama"
version = "0.0.1"
description = ""
authors = ["Pooya Haratian"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/pooyahrtn/RunpodOllama"
repository = "https://github.com/pooyahrtn/RunpodOllama"
documentation = "https://github.com/pooyahrtn/RunpodOllama"


[tool.poetry.dependencies]
python = ">=3.9,<3.13"
flask = "^3.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
litellm = { extras = ["proxy"], version = "^1.20.2" }
types-requests = "^2.31.0.20240125"
typer = "^0.9.0"
runpod = "^1.5.3"
rich = "^13.7.0"
inquirer = "^3.2.3"

# Pre-specify binary packages for better compatibility with Python 3.12
aiohttp = "^3.9.1"  # This version has wheels for Python 3.12


[tool.poetry.group.examples.dependencies]
openai = "^1.10.0"
langchain-community = "^0.0.16"
langchain = "^0.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
local_proxy = 'runpod_ollama:run_local_proxy'
runpod-ollama = 'runpod_ollama:run_cli'
